requirements.txt
streamlit
google-generativeai
streamlit-option-menu
speechrecognition
pyaudio
